+++ libs/@guardian/source-foundations/src/mq/mq.test.ts
@@ -11,7 +11,9 @@ it('from should return a min-width media query', () => {
 it('until should return a max-width media query', () => {
 	const max: Breakpoint = 'wide';
 
-	expect(until[max]).toBe(`@media (max-width: ${`${breakpoints[max] - 1}px`})`);
+	expect(until[max]).toBe(
+		`@media (max-width: ${`${breakpoints[max] - 0.1}px`})`,
+	);
 });
 
 it('between should return a media query with min- and max-width', () => {
@@ -20,7 +22,7 @@ it('between should return a media query with min- and max-width', () => {
 
 	expect(between[min].and[max]).toBe(
 		`@media (min-width: ${`${breakpoints[min]}px`}) and (max-width: ${`${
-			breakpoints[max] - 1
+			breakpoints[max] - 0.1
 		}px`})`,
 	);
 });