+++ libs/@guardian/source-react-components/src/user-feedback/InlineSuccess.stories.tsx
@@ -6,6 +6,9 @@ import type { UserFeedbackProps } from './types';
 const meta: Meta<typeof InlineSuccess> = {
 	title: 'InlineSuccess',
 	component: InlineSuccess,
+	args: {
+		size: 'medium',
+	},
 };
 
 export default meta;
@@ -22,6 +25,28 @@ export const InlineSuccessDefaultTheme: StoryFn<typeof InlineSuccess> =
 export const InlineSuccessBrandTheme: StoryFn<typeof InlineSuccess> =
 	Template.bind({});
 InlineSuccessBrandTheme.parameters = {
+	backgrounds: {
+		default: 'brandBackground.primary',
+	},
+	theme: userFeedbackThemeBrand,
+};
+
+// *****************************************************************************
+
+export const InlineSuccessSmallDefaultTheme: StoryFn<typeof InlineSuccess> =
+	Template.bind({});
+InlineSuccessSmallDefaultTheme.args = {
+	size: 'small',
+};
+
+// *****************************************************************************
+
+export const InlineSuccessSmallBrandTheme: StoryFn<typeof InlineSuccess> =
+	Template.bind({});
+InlineSuccessSmallBrandTheme.args = {
+	size: 'small',
+};
+InlineSuccessSmallBrandTheme.parameters = {
 	backgrounds: {
 		default: 'brandBackground.primary',
 	},