+++ libs/@guardian/source/src/components/react/index.ts
@@ -1,0 +1,1 @@
+export * from '@guardian/source-react-components';