+++ tools/scripts/maintain-readme/index.mjs
@@ -30,6 +30,7 @@ const nxTargets = await getNxTargets();
 
 const nxTargetsList = [
 	"You can also run individual project's Nx targets by running `make <target>`. <details><summary>Nx targets</summary>",
+	'',
 ];
 for (const target of nxTargets) {
 	nxTargetsList.push(`- \`make ${target}\``);